feat(notification-permission): reworked hook
